feat: task-1 according to the module #1
- refactoring code
