--- src/app/counters/counter-comments/counter-comments.component.ts
@@ -1,16 +1,0 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
-
-@Component({
-  selector: 'app-counters-comments',
-  templateUrl: './counter-comments.component.html',
-  styleUrls: ['./counter-comments.component.scss'],
-})
-export class CounterCommentsComponent {
-  @Input() public counter!: string;
-
-  @Output() public comments: EventEmitter<string> = new EventEmitter();
-
-  onComments() {
-    this.comments.emit(this.counter);
-  }
-}
