--- src/app/counters/counter-dislike/counter-dislike.component.ts
@@ -1,16 +1,0 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
-
-@Component({
-  selector: 'app-counters-dislike',
-  templateUrl: './counter-dislike.component.html',
-  styleUrls: ['./counter-dislike.component.scss'],
-})
-export class CounterDislikeComponent {
-  @Input() public counter!: string;
-
-  @Output() public dislike: EventEmitter<string> = new EventEmitter();
-
-  onDislike() {
-    this.dislike.emit(this.counter);
-  }
-}
