--- src/app/elements/button/button.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-button-component',
-  templateUrl: './button.component.html',
-  styleUrls: ['./button.component.scss'],
-})
-export class ButtonComponent {}
