--- src/app/header/logo/logo.component.scss
@@ -1,4 +1,0 @@
-.app-logo {
-  width: 50px;
-  height: 50px;
-}
