--- src/app/header/settings-button/settings-button.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SettingsButtonComponent } from './settings-button.component';
-
-describe('SettingsButtonComponent', () => {
-  let component: SettingsButtonComponent;
-  let fixture: ComponentFixture<SettingsButtonComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ SettingsButtonComponent ],
-    })
-      .compileComponents();
-
-    fixture = TestBed.createComponent(SettingsButtonComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
