--- src/app/header/settings-button/settings-button.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-settings-button',
-  templateUrl: './settings-button.component.html',
-  styleUrls: ['./settings-button.component.scss'],
-})
-export class SettingsButtonComponent {
-  handleClick() {
-    console.log('handleClick');
-  }
-}
