+++ src/app/search/search-item/search-item.component.html
@@ -1,0 +1,5 @@
+<div>
+    <div class="flex justify-center items-center">
+        Search item
+    </div>
+</div>
