+++ src/app/search/search-item/search-item.component.scss
@@ -1,0 +1,8 @@
+.header {
+  background: gray;
+  padding: 20px;
+
+  &-search {
+
+  }
+}
