+++ src/app/search/search-item/search-item.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-search-item-component',
+  templateUrl: './search-item.component.html',
+  styleUrls: ['./search-item.component.scss'],
+})
+export class SearchItemComponent {}
