+++ src/app/search/search-response.model.ts
@@ -1,0 +1,12 @@
+import IPost from "./search-item.model";
+
+interface IPageInfo {
+    totalResults: number,
+    resultsPerPage: number,
+}
+export default interface ISearchResponse {
+    kind: string,
+    etag: string,
+    pageInfo: IPageInfo,
+    items: IPost[],
+}
