+++ src/app/search/search-results/search-results.component.scss
@@ -1,0 +1,8 @@
+.header {
+  background: gray;
+  padding: 20px;
+
+  &-search {
+
+  }
+}
