+++ src/app/search/search-results/search-results.component.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+import { SearchResultsComponent } from './search-results.component';
+
+describe('-SearchItemComponent', () => {
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [
+        RouterTestingModule,
+      ],
+      declarations: [
+        SearchResultsComponent,
+      ],
+    }).compileComponents();
+  });
+});
