--- src/app/video/video-item/video-item.component.html
@@ -1,22 +1,0 @@
-<div class="max-w-sm p-6 bg-white border border-gray-200 rounded-lg shadow dark:bg-gray-800 dark:border-gray-700">
-    <img [src]="item.snippet.thumbnails.medium.url" class="mb-6" alt="image"/>
-    <div class="flex space-x-3">
-        <div>
-            <app-counters-like [counter]="item.statistics.likeCount" (click)="onLike()"></app-counters-like>
-        </div>
-        <div>
-            <app-counters-dislike [counter]="item.statistics.dislikeCount" (click)="onDislike()"></app-counters-dislike>
-        </div>
-        <div>
-            <app-counters-views [counter]="item.statistics.viewCount" (click)="onViews()"></app-counters-views>
-        </div>
-        <div>
-            <app-counters-comments [counter]="item.statistics.commentCount" (click)="onComments()"></app-counters-comments>
-        </div>
-    </div>
-    <h5 class="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">{{ item.snippet.title }}</h5>
-    <button type="button" disabled
-            class="inline-flex items-center px-3 py-2 text-sm font-medium text-center text-white bg-blue-700 rounded-lg hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
-        more
-    </button>
-</div>
