--- src/app/video/video-list/video-list.component.html
@@ -1,11 +1,0 @@
-<h1 class="my-10 text-4xl text-center font-extrabold leading-none tracking-tight text-gray-900 md:text-5xl lg:text-6xl dark:text-white">My posts</h1>
-<div class="flex flex-wrap flex-gap-1">
-    <app-video-item
-            *ngFor="let item of list"
-            [item]="item"
-            (like)="handleClickLike(item)"
-            (dislike)="handleClickDislike(item)"
-            (views)="handleClickViews(item)"
-            (comments)="handleClickComments(item)"
-    ></app-video-item>
-</div>
